# manualDevelopment/backend/docs/swagger.yaml
openapi: 3.0.0
info:
  title: School Equipment Lending Portal API
  description: |

    ## Authentication
    All endpoints (except login and setup-password) require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```

    ## How to Authenticate in Swagger UI
    1. Click the **Authorize** button (top right)
    2. Login using POST /auth/login to get a token
    3. Copy the token from response
    4. In the Authorize dialog, paste the token (without "Bearer " prefix)
    5. Click Authorize

    ## User Roles
    - **Admin**: Full access to all resources
    - **Teacher**: Can approve/deny requests, view statistics
    - **Student**: Can request equipment, view own requests

    ## Base URL
    Replace `{BASE_URL}` with your API endpoint (e.g., `http://localhost:4000`)
  version: 1.0.0
  contact:
    name: API Support
    email: support@schoolequipment.com
  license:
    name: MIT
servers:
  - url: http://localhost:4000/api
    description: Local server
  - url: https://school-equipment-landing-portal-q6q.vercel.app/api
    description: Production server (Vercel Live)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          description: Unique user identifier
          example: 1
        email:
          type: string
          format: email
          description: User email address
          example: john@school.com
        full_name:
          type: string
          description: User's full name
          example: John Doe
        role:
          type: string
          enum: [Admin, Teacher, Student]
          description: User role
          example: Admin
        phone:
          type: string
          description: User's phone number
          example: +1234567890
        department:
          type: string
          description: User's department
          example: Science
        status:
          type: string
          enum: [Active, Inactive, Pending Setup]
          description: User account status
          example: Active
        created_at:
          type: string
          format: date-time
          description: Account creation timestamp
          example: 2025-01-15T10:30:00Z
        updated_at:
          type: string
          format: date-time
          description: Account last update timestamp
          example: 2025-01-15T10:30:00Z
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email
          example: user@school.com
        password:
          type: string
          description: User password (minimum 6 characters)
          minLength: 6
          example: password123
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Login successful
        data:
          type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            token:
              type: string
              description: JWT token for authentication
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    SetupPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          description: Password setup token from email
          example: setup_token_received_via_email
        password:
          type: string
          minLength: 6
          description: New password (minimum 6 characters)
          example: newPassword123

    # Pagination Schemas
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message describing what went wrong
        errors:
          type: array
          items:
            type: object
            properties:
              msg:
                type: string
              param:
                type: string
              value:
                type: string
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticate user with email and password. Returns JWT token for subsequent requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              student:
                value:
                  email: student@school.com
                  password: password123
              teacher:
                value:
                  email: teacher@school.com
                  password: password123
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
              examples:
                success:
                  value:
                    success: true
                    message: Login successful
                    data:
                      user:
                        id: 1
                        email: user@school.com
                        full_name: John Doe
                        role: Teacher
                        status: Active
                      token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidPassword:
                  value:
                    success: false
                    message: Invalid email or password
                pendingSetup:
                  value:
                    success: false
                    message: Please complete your account setup first. Check your email for the setup link.
                inactive:
                  value:
                    success: false
                    message: Your account has been deactivated. Please contact the administrator.
  /auth/setup-password:
    post:
      tags:
        - Authentication
      summary: Setup Password (New User)
      description: Set password for new users via email setup link. This endpoint is called when users receive their account setup email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupPasswordRequest"
      responses:
        "200":
          description: Password setup successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password setup successful. You can now login.
        "400":
          description: Invalid request or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get Current User
      description: Get information about the currently authenticated user
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user information
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== USER MANAGEMENT ENDPOINTS ====================
...